+++ app/views/account/announces/_form.html.erb
@@ -1,18 +1,13 @@
 <div class="container padded">
   <div class="row">
     <div class="col-sm-6 col-sm-offset-3">
-      <%= simple_form_for @announce do |f| %>
+      <%= simple_form_for(@announce, url: @announce.new_record? ? account_announces_path : account_announce_path(@announce)) do |f| %>
       <%= f.association :company %>
       <%= f.input :title %>
       <%= f.input :price %>
       <%= f.input :kind %>
       <%= f.input :number_of_share %>
       <%= f.input :reason %>
-
-
-
-
-
       <%= f.submit %>
       <% end %>
     </div>
