+++ app/views/account/announces/show.html.erb
@@ -1,0 +1,11 @@
+<h1> Nouvelle annonce </h1>
+<p><%= @announce.title %></p>
+<p><%= @announce.price %></p>
+<p><%= @announce.kind %></p>
+<p><%= @announce.number_of_share %></p>
+
+
+<%= link_to "Back", account_announces_path %>
+<%= link_to "Delete", account_announce_path, method: :delete, data: {confirm: "Are you sure ?"}, class: 'btn btn-default' %>
+<%= link_to "Edit", edit_account_announce_path %>
+
