+++ app/views/announces/show.html.erb
@@ -6,6 +6,3 @@
 
 
 <%= link_to "Back", root_path %>
-<%= link_to "Delete", announce_path, method: :delete, data: {confirm: "Are you sure ?"}, class: 'btn btn-default' %>
-<%= link_to "Edit", edit_announce_path %>
-
