+++ app/views/pages/home.html.erb
@@ -24,7 +24,7 @@
         available without you adding any controller or touching the <code>config/routes.rb</code> file!
       </p>
 
-      <p><%= link_to "Ajouter une annonce", new_announce_path %></p>
+      <p><%= link_to "Ajouter une annonce", new_account_announce_path %></p>
       <p><%= link_to "Voir toutes les annonces", announces_path %></p>
       <li>
         <%= link_to "Connexion", new_user_session_path %>
