+++ config/routes.rb
@@ -4,6 +4,9 @@ Rails.application.routes.draw do
   ActiveAdmin.routes(self)
   devise_for :users, controllers: { omniauth_callbacks: "users/omniauth_callbacks" }
 
+  namespace :account do
+    resources :announces
+  end
 
   resources :announces do
     resources :purchases, only: [:new, :create]
